Moved threshold tolerance issue from pack test to golden generator.
