+++ tt_metal/tt-llk/tests/python_tests/quasar/test_eltwise_binary_broadcast_quasar.py
@@ -168,5 +168,5 @@ def test_eltwise_binary_broadcast_quasar(
     res_tensor = torch.tensor(res_from_L1, dtype=torch_format)
 
     assert passed_test(
-        golden_tensor, res_tensor, formats.output_format, print_errors=True
+        golden_tensor, res_tensor, formats.output_format
     ), "Assert against golden failed"
